Woohoo physics kinda done
- Subframes implemented
- Having a collision no longer short-curciuts the distance travelled for that tick
- Multiple collisions per tick supported
- Vertex collisions are functional
- Colliding w/ multiple walls is now handled correctly

Everything functions with any sized radius or any amount of walls*. There are
of course limitation with precision, but for the most part this all
works very well.

* - Collision algorithm is O(n) w/ respsect to number of walls and most frames
it's the most expensive computation.

Talking about performance, as of now the examples I've tested with have been small.
However, they're has been no evident lag spike so it should be fine. When
performance becomes an issue the physics can be rewritten for a different
collision mechanism and the play field should be spatially partioned
to reduce the # of collision checks.